requirements.txt
dynaconf==3.1.7
flask==2.0.2
flask-cors==3.0.10
flask-pydantic-spec==0.2.1
flask-restful==0.3.9
flask-sqlalchemy==2.5.1
huey==2.4.3
pydantic==1.9.0
toml==0.10.2
